double -> float

+++ src/camera.h
@@ -15,34 +15,34 @@ class Camera {
 
   void LoadMatrices(Shader *shader) const;
 
-  inline void RelativeRotate(const glm::dvec3 amount) { rotation_ += amount; }
-  inline void RelativeMove(const glm::dvec3 amount) {
+  inline void RelativeRotate(const glm::vec3 amount) { rotation_ += amount; }
+  inline void RelativeMove(const glm::vec3 amount) {
     position_ += amount.x * look_vector() + amount.y * normal_vector() +
                  amount.z * up_vector();
   }
-  inline void AbsoluteMove(const glm::dvec3 amount) { position_ += amount; }
+  inline void AbsoluteMove(const glm::vec3 amount) { position_ += amount; }
 
   inline void set_aspect(const int width, const int height) {
     aspect_ = static_cast<double>(width) / height;
   }
 
  private:
-  inline glm::dvec3 look_vector() const {
-    return glm::rotateZ(glm::rotateY(glm::dvec3(1., 0., 0.), rotation_.y),
+  inline glm::vec3 look_vector() const {
+    return glm::rotateZ(glm::rotateY(glm::vec3(1., 0., 0.), rotation_.y),
                         rotation_.z);
   }
-  inline glm::dvec3 up_vector() const {
-    return glm::rotateZ(glm::rotateY(glm::dvec3(0., 0., 1.), rotation_.y),
+  inline glm::vec3 up_vector() const {
+    return glm::rotateZ(glm::rotateY(glm::vec3(0., 0., 1.), rotation_.y),
                         rotation_.z);
   }
-  inline glm::dvec3 normal_vector() const {
+  inline glm::vec3 normal_vector() const {
     return glm::cross(up_vector(), look_vector());
   }
 
-  glm::dvec3 position_{0., 0., 2.};
+  glm::vec3 position_{0., 0., 2.};
   // This vector is not physically representative of the camera's rotation,
   // rather the x component controls the roll, the y component controls the
   // pitch, and the z component controls the yaw.
-  glm::dvec3 rotation_{0., glm::half_pi<double>(), 0.};
+  glm::vec3 rotation_{0., glm::half_pi<double>(), 0.};
   double aspect_{1.};
 };
